refactor: simplifica deploy usando zero-config NestJS do Vercel
Remove vercel-handler.ts e serverless-http: Vercel detecta
automaticamente NestJS via src/main.ts (zero-config).

Remove bootstrap-app.ts (main.ts volta a ser arquivo único).

Remove apps/api/vercel.json: sem necessidade com zero-config.

Adiciona postinstall prisma generate para garantir cliente
atualizado no build do Vercel (recomendação oficial Prisma).

Mantém apps/web/vercel.json (necessário para SPA rewrites).
